+++ lib/legion/api.rb
@@ -28,7 +28,6 @@ require_relative 'api/rbac'
 require_relative 'api/auth'
 require_relative 'api/auth_worker'
 require_relative 'api/auth_human'
-require_relative 'api/auth_kerberos'
 require_relative 'api/capacity'
 require_relative 'api/audit'
 require_relative 'api/metrics'
@@ -104,7 +103,6 @@ module Legion
     register Routes::Auth
     register Routes::AuthWorker
     register Routes::AuthHuman
-    register Routes::AuthKerberos
     register Routes::Capacity
     register Routes::Audit
     register Routes::Metrics